+++ WorkFlowWebServices/src/main/java/com/gs/photos/repositories/impl/IHbaseRatingsDAO.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 import com.gs.photo.workflow.hbase.dao.IHbaseMetaDataDAO;
 import com.workflow.model.HbaseRatings;
 
-public interface IHbaseRatingsDAO extends IHbaseMetaDataDAO<HbaseRatings, Integer> {
+public interface IHbaseRatingsDAO extends IHbaseMetaDataDAO<HbaseRatings, Long> {
 
     void flush() throws IOException;
 }
